+++ src/Models/JoystickState.cs
@@ -7,12 +7,14 @@ namespace AnalogtoKey.Models
         public string DeviceGuid { get; set; } = "";
         public string Name { get; set; } = "";
 
-        public int AxisX { get; set; }
-        public int AxisY { get; set; }
-        public int AxisZ { get; set; }
+        public int AxisX  { get; set; }
+        public int AxisY  { get; set; }
+        public int AxisZ  { get; set; }
         public int AxisRx { get; set; }
         public int AxisRy { get; set; }
         public int AxisRz { get; set; }
+        public int Slider0 { get; set; }
+        public int Slider1 { get; set; }
 
         public int HatSwitch { get; set; }
 
@@ -22,13 +24,15 @@ namespace AnalogtoKey.Models
 
         public int GetAxis(string name) => name switch
         {
-            "AxisX"  => AxisX,
-            "AxisY"  => AxisY,
-            "AxisZ"  => AxisZ,
-            "AxisRx" => AxisRx,
-            "AxisRy" => AxisRy,
-            "AxisRz" => AxisRz,
-            _        => 0
+            "AxisX"   => AxisX,
+            "AxisY"   => AxisY,
+            "AxisZ"   => AxisZ,
+            "AxisRx"  => AxisRx,
+            "AxisRy"  => AxisRy,
+            "AxisRz"  => AxisRz,
+            "Slider0" => Slider0,
+            "Slider1" => Slider1,
+            _         => 0
         };
     }
 }